+++ apiWithDb/src/main/java/com/example/apiWithDb/service/CompanyService.java
@@ -9,8 +9,8 @@ public interface CompanyService {
 
     public String createCompany(Company company,Authentication authentication);
     public String updateCompany(Company company,Authentication authentication);
-    public String deleteCompany(Integer companyId,Authentication authentication);
-    public Company getCompanyForAdmin(Integer companyId,Authentication authentication);
+    public String deleteCompany(Long companyId,Authentication authentication);
+    public Company getCompanyForAdmin(Long companyId,Authentication authentication);
     public Company getCompanyForEmployee(Authentication authentication);
     default Company getCompany(String CompanyId) {
         return null;
